Usage Description Groups
Added new usage_description type used as a header for usage description
groups. Also added utils.c, which will be used for special ubiquitous
utilities.

In utils, I added a BUG function, which is used a bit like an assertion.
This is very similar to what is implemented in git-core.