Improvements to run-command API
Improvements to make the run-command api to be more useful and robust.
One such improvement is to make run-command use the execve() system call
rather than popen(), which provides very little functionality.

Fixed a bug in argv-array that caused argv_array_collapse() to fail on
empty arrays. Added unit tests.

Added a str_array_sort() function that sorts the array in strcmp()
order. Makes use of qsort() in stdlib.

Renamed main.c to git-chat.c. Updated CMakeLists.txt to reflect this
change. A reload and rebuild of the project will be necessary.