Implemented run-command Plumbing
run-command now uses execve() interally, rather than popen(). It has
also been adapted to accept an alternate directory from which the
process will be executed.

If execve fails to run the program directly, a second attempt is made by
running the program through 'sh -c'.

Left to implement is the use of pipes to read/write to/from the parent
process to the child process.

Various minor improvements.