run-command: minor tweaks to design
Changed the implementation of run-command such that all standard streams
are inherited from the parent, unless capturing the output. This ensures
that the child process can print stdout and stderr in order, which is
far simplier.

Added documentation to reflect this change.

In the future, if we find ourselves restricted by the current
implementation, we can extend it further.