run-command: bug fixes
Fixed a bug that was causing the tests to fail. merge_env() accepts an
empty str_array as the resultant env, but we were populating it with the
delta_env before passing it to merge_env(), so we ended up with
duplicate environment variables. On macos, printenv only returns the first
environent variable that matches the argument, while linux prints all.