git-chat extension
If git-chat was invoked with a subcommand that doens't exist in the list
of builtins, attempts to run the subcommand as an extension. Similar to
git and their extension model, search the PATH for any executables named
git-chat-<extension name>, and execute it.

Also moved a few functions to fs-utils