Restructure CMakeLists and Add Integration Tests
Restructured the CMakeLists build files such that the test directory is
a project subdirectory. This helps keep the test build files together in
a single location within the build directory, and simplifies the
CMakeLists.

Changes to test sources and resources will now cause the updated files
to be built before the target is executed. This means that changing a
resource file will not require the entire cmake cache to be rebuilt.

Also added a simple integration test. In the future, these will be more
robust and will test git-chat from end to end.

Modifications to travis scripts were made to support the new build
structure.

Update project README.