Code Cleanup and Improvements
Various minor improvements to the codebase. The most notable changes are
the following:
- the sentinel function attribute was added to functions that expect
variadic arguments with a NULL pointer terminating the args. This will
help reduce developer error.
- renamed various usage strings
- ensure that fs-utils is_executable is idempotent on the errno.
- renamed file pointers from fd to fp.
- cmake now runs the integration test suite in bash, rather than sh.
This is needed, since sh is not a platform agnostic way of running
scripts.
- Fixed bug around strbuf_attach that would cause it to fail when
attaching large strings, since the buffer was not growing correctly.
- Added const keyword where appropriate.