Enable Valgrind Memcheck on Integration Test Suite
Updated integration-runner with mechanism for running integration tests
under valgrind memcheck. This is added as an option to the integration
runner script.

If enabled, git-chat is wrapped by a script valgrind.sh. The script is
symbollically linked and added to the PATH so that it appears to be a
normal git-chat binary.