Improvements to run-command, Tests and Utils
Improvements were made to run-command to ensure that the caller is able to
read and write to and from standard streams such to avoid filling the
pipes and blocking. In doing so, improved the ability to chain child
process streams together.

In utils, added two new functions recoverable_write() and
recoverable_read(), which are more resilient to recoverable errors.

In strbuf, added a new function to split a string buffer by delimiter.

In str-array, added a function to clear the contents of the array
without fully releasing the array resources.

Integration tests now write verbose output to a file, and travis is now
instructed to output the contents of the log file on a test failure.