str-array improvements
The str-array api is a very useful utility for manipulating string data.
However, the api is limited in its usefulness, given that arbitrary data
could not be attached to each string entry.

Now, the str-array api has been extended such that each string entry can
also reference an arbitrary memory location. This is useful as an
associative array, or simple map, between string data and other types of
data.

It is also useful for dynamically building arrays of data.

The str_array_detach() functions were also tweaked to always return an
array that is, at least, one larger than the overall length, and the last entry
is NULL. This is useful when library functions accept a null-terminated
array.