Add Dependency on GPGME Library
After much trial and error using gpg through the run-command api, it was
found to be far too brittle and difficult to maintain. So, to avoid a messy
interface between git-chat and gpg, use the GPGME library for running gpg
commands.

A CMake script to find GPGME was adapted for use in the git-chat
project. It is run at build time to ensure that gpgme 'vanilla' is
installed and is usable.

Created a gpg-interface api for interfacing with gpgme. It provides
functions for rebuilding the keyring, fetching gpg keys, filtering and
manipulating gpg key lists, encrypting plaintext messages, and so on.

Updated git-chat.c to print the GPGME library version number, along with
git-chat and git version numbers.

Updated project README.