Changes to Accomodate GPGME in Travis CI
Travis by default uses an older version of Ubuntu, which using apt can
only install gpgme version 1.6.0, which is too out of date. Updated the
travis build file to default to Ubuntu Bionic, which has support for
gpgme 1.10.0.

Since gpg2 is not available on macos, and is needed by the integeration
test suite, updated the integration runner to create a symlink to gpg.