git-chat-message with Symmetrically Encrypted Messages
The symmetric message encrypt feature was added, which came along with
various refactors and improvements. Added a --passphrase option which
can be used to bypass the normal pinentry method. This is particularily
useful in scripts that may run git-chat.

safe_create_dir() was taught to respect the source file modes.

Added valgrind suppression file which is used to suppress the gpgme
memory warnings.