cmake fail-fast when GPGME cannot be found
If the GPGME library cannot be found during the build process, we should
terminate with an appropriate message. This was not done before, and if
GPGME was missing, the build would succeed but the compilation would
fail. Since cmake knows if the library could not be found, this way we
can display a much friendlier error message.

Furthermore, we weren't adding the include directories for GPGME. This
would cause compilation to fail on systems where GPGME is not installed
in a traditional location (in Homebrew keg, for instance).