created new parse-options api to replace usage api
The old 'usage' api served it's purpose, but was not very
developer-friendly. For once, each argument needed to be parsed manually
within each builtin, and the code to do so was monsterous and
error-prone. To avoid this, the parse-options api was created which
handles parsing the arguments.

The usage_description struct was renamed to usage_string, and the
option_description struct was renamed to command_option.

Preprocessor macros for building command_option[] array is largely the
same as before, except now some macros accept a pointer which is set by
parse_options if an argument matches an option_description. For
instance, if an option_description is of type OPT_BOOL and a pointer to
an int is provided, then if the user supplies the boolean option, the
integer at the address is set to 1.

Added tests for the new api, and made minor changes to related tests.