unit test-lib improvements
The unit test library was improved. Prior to these changes, the test
library structure was a bit inhibitive, where using asserts within
iterative constructs in the test body would break from the inner loop,
and not jump to the test teardown section.

By using goto's within the asserts, we avoid this problem.