Implement strbuf_clear() Function
Created a simple function that can be used to clear the content of a
strbuf. This is particularly useful for reusing allocated buffers for a
new purpose without the overhead associated with allocating and freeing
new memory.

For instance, in a loop it is no longer necessary to release the strbuf
and then reinitialize it when it is needed more than once.

Added a unit test to cover this new function.