Improvements to Unit Test Library
No that the unit test library is using goto's instead of 'break', we can
now wrap the assertions in a do while() to make sure that the semicolon
is necessary at the end of the assertion.

This also ensures that we stay consistent throughout all the tests, but
also avoid compiler warnings with unecessary semicolons at the end of
each assertion.