Reimplemented parse-config API
The existing parse-config api wsas written in the early stages of this
project, and was needlessly restrictive. The new implementation is more
modular and easier to manipulate.

The config file data structure now uses a linked list internally, which
makes it more efficient to add and remove entries. The entry structure
was also made opaque, requiring functions to accessing string data
within. This helps separate the concerns of the caller from the
internals of parse-config.

Tests were rewritten to appropriately test the new changes.

git-chat-init was modified to use the new api.