git-chat-message bug fix
When creating a new message without supplying the body of the message
with the -m switch, the command would fail with a fatal error if the
GC_EDITMSG file is missing.

To work around this, added the O_CREAT flag when opening the file, which
will create the file if missing.