git-chat-read improvements
A few improvements were made to git-chat-read under this revision to
improve user experience. Message headings in the output is now shown in
color, unless the --no-color option is provided. The --max-count option
was added to limit the number of messages shown. The semantics of the
pager was improved for easier reading.