git-chat-message: add --reply and --compose
Introduced new --reply and --compose options to show decrypted messages
in a vim split window. This makes it easier to respond to messages.

Messages are decrypted and shown in the split window much in the same
way that messages are decrypted and shown in git-chat-read. To make this
work, the commit traversal machinery was pulled out of the git-chat-read
builtin and into it's own component, graph-traversal.